+++ CHANGES.rst
@@ -8,6 +8,7 @@ Changelog
 - added `metadata-from-name` filter to extract meta-data from the current input file name
 - requiring seppl>=0.1.3 now
 - added `inspect` filter that allows inspecting data interactively as it passes through the pipeline
+- added `empty_str_if_none` helper method to `ldc.text_utils` to ensure no None/null values are output with writers
 
 
 0.2.0 (2024-02-27)